move cullable to utils
