+++ src/ts/mandelbulb/mandelbulb.ts
@@ -25,7 +25,7 @@ import { Scene } from "@babylonjs/core/scene";
 import { OrbitProperties } from "../orbit/orbitProperties";
 import { RingsUniforms } from "../rings/ringsUniform";
 import { Vector3 } from "@babylonjs/core/Maths/math.vector";
-import { Cullable } from "../bodies/cullable";
+import { Cullable } from "../utils/cullable";
 import { OrbitalObjectPhysicalProperties } from "../architecture/physicalProperties";
 import i18n from "../i18n";
