+++ src/ts/utils/cullable.ts
